more platform agnostic

+++ src/point.rs
@@ -2,11 +2,11 @@ use crate::streamed_data::StreamedData;
 use crate::streamed_data::StreamedState;
 
 #[derive(Ord, PartialOrd, Eq, PartialEq, Debug, Copy, Clone)]
-pub struct Timeval(pub i64, pub i64);
+pub struct Timeval(pub isize, pub isize);
 
 impl Timeval {
     pub fn from_timeval(t: ::libc::timeval) -> Timeval {
-        Timeval(t.tv_sec, t.tv_usec)
+        Timeval(t.tv_sec as isize, t.tv_usec as isize)
     }
 }
